refactor: repurpose the repo to clarity-vscode
