--- .cargo/config
@@ -1,5 +1,0 @@
-[alias]
-xtask = "run --package xtask --bin xtask --"
-install-ra = "run --package xtask --bin xtask -- install" # for backwards compat
-tq = "test -- -q"
-qt = "tq"
