+++ .eslintignore
@@ -6,3 +6,4 @@ dist
 coverage
 # don't lint out directory
 out
+
