+++ .github/workflows/ci.yaml
@@ -25,177 +25,6 @@ jobs:
         with:
           access_token: ${{ github.token }}
 
-  audit:
-    name: Audit and format
-    runs-on: ubuntu-latest
-    needs: pre_run
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v2
-
-      - name: Install Rust toolchain
-        uses: actions-rs/toolchain@v1
-        with:
-          toolchain: stable
-          profile: minimal
-          components: rustfmt
-          override: true
-
-      - name: Set Cargo file permissions
-        run: sudo chown -R $(whoami):$(id -ng) ~/.cargo/
-
-      - name: Cache cargo
-        uses: actions/cache@v2
-        with:
-          path: ~/.cargo/
-          key: ${{ runner.os }}-cargo-${{ hashFiles('**/Cargo.lock') }}
-
-      - name: Install dependencies
-        run: cargo install cargo-audit
-
-      - name: Run audit
-        run: cargo audit --ignore RUSTSEC-2021-0076 --ignore RUSTSEC-2018-0015 --ignore RUSTSEC-2021-0124 --ignore RUSTSEC-2021-0119 --ignore RUSTSEC-2022-0013
-
-      - name: Run rustfmt
-        run: cargo fmt --all -- --check
-
-  test:
-    name: Run tests
-    runs-on: ubuntu-latest
-    needs: pre_run
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v2
-
-      - name: Install Rust toolchain
-        uses: actions-rs/toolchain@v1
-        with:
-          toolchain: nightly
-          profile: minimal
-          components: llvm-tools-preview
-          override: true
-
-      - name: Cache cargo
-        uses: actions/cache@v2
-        id: cache-cargo
-        with:
-          path: |
-            ~/.cargo/bin/
-            ~/.cargo/registry/index/
-            ~/.cargo/registry/cache/
-            ~/.cargo/git/db/
-            target/
-          key: ${{ runner.os }}-cargo-${{ hashFiles('**/Cargo.lock') }}
-
-      - name: Install dependencies
-        if: steps.cache-cargo.outputs.cache-hit != 'true'
-        run: cargo install grcov
-
-      - name: Unit Tests
-        env:
-          RUSTFLAGS: "-Zinstrument-coverage"
-          LLVM_PROFILE_FILE: "clarity-lsp-%p-%m.profraw"
-        run: cargo build && cargo test
-
-      - name: Generate coverage
-        run: grcov . --binary-path ./target/debug/ -s . -t lcov --branch --ignore-not-existing --ignore "/*" -o lcov.info
-
-      - name: Upload coverage report
-        uses: codecov/codecov-action@v1
-        with:
-          flags: unittests
-          name: clarity-lsp
-          verbose: true
-
-  dist:
-    name: Build distributions
-    runs-on: ${{ matrix.os }}
-    needs: pre_run
-    strategy:
-      fail-fast: false
-      matrix:
-        include:
-          - os: ubuntu-latest
-            platform: linux-x64
-            target: x86_64-unknown-linux-gnu
-            linker_package:
-            cc: ""
-          - os: windows-latest
-            platform: windows-x64
-            target: x86_64-pc-windows-msvc
-            linker_package:
-            cc: ""
-          - os: macos-latest
-            platform: macos-x64
-            target: x86_64-apple-darwin
-            linker_package:
-            cc: ""
-
-    steps:
-      - name: Configure git to use LF (Windows)
-        if: matrix.os == 'windows-latest'
-        run: |
-          git config --global core.autocrlf false
-          git config --global core.eol lf
-
-      - name: Checkout repository
-        uses: actions/checkout@v2
-
-      - name: Install Rust toolchain
-        uses: actions-rs/toolchain@v1
-        with:
-          toolchain: stable
-          target: ${{ matrix.target }}
-          profile: minimal
-          override: true
-
-      - if: matrix.os == 'ubuntu-latest'
-        run: sudo chown -R $(whoami):$(id -ng) ~/.cargo/
-
-      - name: Cache cargo
-        uses: actions/cache@v2
-        with:
-          path: |
-            ~/.cargo/bin/
-            ~/.cargo/registry/index/
-            ~/.cargo/registry/cache/
-            ~/.cargo/git/db/
-            target/${{ matrix.target }}/release/
-          key: ${{ runner.os }}-cargo-${{ hashFiles('**/Cargo.lock') }}
-
-      - name: Build
-        run: cargo build --release --locked --target ${{ matrix.target }}
-
-      - name: Compress artifact (Not Windows)
-        if: matrix.os != 'windows-latest'
-        run: zip --junk-paths ${{ matrix.platform }} ./target/${{ matrix.target }}/release/clarity-lsp
-
-      - name: Compress artifact (Windows)
-        if: matrix.os == 'windows-latest'
-        run: Compress-Archive -Path ./target/${{ matrix.target }}/release/clarity-lsp.exe -DestinationPath ${{ matrix.platform }}
-
-      - name: Upload artifact
-        uses: actions/upload-artifact@v2
-        with:
-          name: clarity-lsp-${{ matrix.platform }}
-          path: ${{ matrix.platform }}.zip
-
-      - name: Unit Tests
-        run: cargo test --release --locked --target ${{ matrix.target }}
-
-      - name: Upload artifact to GH release
-        uses: svenstaro/upload-release-action@v2
-        if: startsWith(github.ref, 'refs/tags/v')
-        with:
-          repo_token: ${{ secrets.GITHUB_TOKEN }}
-          file: ${{ matrix.platform }}.zip
-          asset_name: clarity-lsp-${{ matrix.platform }}.zip
-          tag: ${{ github.ref }}
-
-      # Cleans the `./target` dir after the build such that only dependencies are cached on CI
-      - name: Prepare cache
-        run: cargo xtask pre-cache
-
   typescript:
     name: TypeScript
     runs-on: ubuntu-latest
@@ -250,9 +79,6 @@ jobs:
     runs-on: ubuntu-latest
     if: contains(github.ref, 'main')
     needs:
-      - audit
-      - dist
-      - test
       - typescript
     steps:
       - name: Checkout
