--- .github/workflows/rustdoc.yaml
@@ -1,38 +1,0 @@
-name: rustdoc
-on:
-  push:
-   branches:
-   - main
-
-env:
-  CARGO_INCREMENTAL: 0
-  CARGO_NET_RETRY: 10
-  RUSTFLAGS: -D warnings
-  RUSTUP_MAX_RETRIES: 10
-
-jobs:
-  rustdoc:
-    runs-on: ubuntu-latest
-
-    steps:
-    - name: Checkout repository
-      uses: actions/checkout@v2
-
-    - name: Install Rust toolchain
-      uses: actions-rs/toolchain@v1
-      with:
-        toolchain: stable
-        profile: minimal
-        override: true
-        components: rustfmt, rust-src
-
-    - name: Build Documentation
-      run: cargo doc --all --no-deps
-
-    - name: Deploy Docs
-      uses: peaceiris/actions-gh-pages@364c31d33bb99327c77b3a5438a83a357a6729ad # v3.4.0
-      with:
-        github_token: ${{ secrets.GITHUB_TOKEN }}
-        publish_branch: gh-pages
-        publish_dir: ./target/doc
-        force_orphan: true
