--- Cargo.toml
@@ -1,35 +1,0 @@
-[package]
-name = "clarity-lsp"
-version = "0.7.1"
-description = "Clarity Language Server"
-authors = ["Ludo Galabru <ludovic@galabru.com>"]
-readme = "README.md"
-edition = "2018"
-license = "GPL-3.0-only"
-keywords = ["blockstack", "blockchain", "clarity", "smart-contract", "lsp"]
-exclude = ["vs-client/**"]
-homepage = "https://github.com/hirosystems/clarity-lsp"
-repository = "https://github.com/hirosystems/clarity-lsp"
-categories = ["command-line-utilities", "development-tools", "development-tools::build-utils"]
-
-[workspace]
-members = [".", "xtask/"]
-
-[dependencies]
-tower-lsp = "0.14.0"
-lazy_static = "1.4.0"
-serde_json = "1.0.47"
-serde = "1"
-serde_derive = "1"
-hmac = "=0.10.1"
-pbkdf2 = { version = "=0.7.3", features=["parallel"], default-features = false }
-tokio = { version = "=1.8.1", features = ["rt", "macros", "io-std"] }
-tokio-util = { version = "0.6.5", features = ["codec"] }
-regex = "1.3.4"
-sha2 = "0.9.3"
-toml = { version = "0.5.6", features = ["preserve_order"] }
-tiny-hderive = "=0.3.0"
-bip39 = { version = "=1.0.1", default-features = false }
-libsecp256k1 = "0.5.0"
-# clarity_repl = { package = "clarity-repl", path = "../clarity-repl" }
-clarity_repl = { package = "clarity-repl", version = "=0.15.0", features = ["cli"] }
