--- clarinet-lsp.code-workspace
@@ -1,11 +1,0 @@
-{
-	"folders": [
-		{
-			"path": "."
-		},
-		{
-			"path": "../clarity-repl"
-		}
-	],
-	"settings": {}
-}
