--- editors/code/.gitignore
@@ -1,5 +1,0 @@
-out
-node_modules
-.vscode-test/
-*.vsix
-bundle
