--- editors/code/tslint.json
@@ -1,6 +1,0 @@
-{
-	"rules": {
-		"indent": [true, "tabs"],
-		"semicolon": [true, "always"]
-	}
-}
