+++ src/status_display.ts
