+++ syntaxes/clarity.tmLanguage.json
