--- xtask/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-edition = "2018"
-name = "xtask"
-version = "0.1.0"
-authors = ["Ludo Galabru <ludovic@galabru.com>"]
-publish = false
-
-[lib]
-doctest = false
-
-[dependencies]
-walkdir = "2.3.1"
-pico-args = "0.3.1"
-quote = "1.0.2"
-proc-macro2 = "1.0.8"
-anyhow = "1.0.26"
